refactor!: Refactor models into pipelines
