+++ .gitignore
@@ -140,3 +140,4 @@ cython_debug/
 .vscode
 creds
 .DS_Store
+voltaic-country-280607-ea3eb5348029.json
