+++ caresoft/pipeline/calls.py
@@ -1,0 +1,66 @@
+from caresoft.pipeline.interface import Pipeline
+from caresoft.request_parser import time_request_parser
+
+pipeline = Pipeline(
+    name="Calls",
+    uri="calls",
+    res_fn=lambda x: x["calls"],
+    params_fn=time_request_parser,
+    transform=lambda rows: [
+        {
+            "id": row["id"],
+            "customer_id": row.get("customer_id"),
+            "call_id": row.get("call_id"),
+            "path": row.get("path"),
+            "path_download": row.get("path_download"),
+            "caller": row.get("caller"),
+            "called": row.get("called"),
+            "user_id": row.get("user_id"),
+            "agent_id": row.get("agent_id"),
+            "group_id": row.get("group_id"),
+            "call_type": row.get("call_type"),
+            "start_time": row.get("start_time"),
+            "call_status": row.get("call_status"),
+            "end_time": row.get("end_time"),
+            "wait_time": row.get("wait_time"),
+            "hold_time": row.get("hold_time"),
+            "talk_time": row.get("talk_time"),
+            "end_status": row.get("end_status"),
+            "ticket_id": row.get("ticket_id"),
+            "last_agent_id": row.get("last_agent_id"),
+            "last_user_id": row.get("last_user_id"),
+            "call_survey": row.get("call_survey"),
+            "call_survey_result": row.get("call_survey_result"),
+            "missed_reason": row.get("missed_reason"),
+        }
+        for row in rows
+    ],
+    schema=[
+        {"name": "id", "type": "INTEGER"},
+        {"name": "start_time", "type": "TIMESTAMP"},
+        {"name": "customer_id", "type": "INTEGER"},
+        {"name": "call_id", "type": "STRING"},
+        {"name": "path", "type": "STRING"},
+        {"name": "path_download", "type": "STRING"},
+        {"name": "caller", "type": "STRING"},
+        {"name": "called", "type": "STRING"},
+        {"name": "user_id", "type": "STRING"},
+        {"name": "agent_id", "type": "STRING"},
+        {"name": "group_id", "type": "INTEGER"},
+        {"name": "call_type", "type": "INTEGER"},
+        {"name": "call_status", "type": "STRING"},
+        {"name": "end_time", "type": "TIMESTAMP"},
+        {"name": "wait_time", "type": "STRING"},
+        {"name": "hold_time", "type": "STRING"},
+        {"name": "talk_time", "type": "STRING"},
+        {"name": "end_status", "type": "STRING"},
+        {"name": "ticket_id", "type": "INTEGER"},
+        {"name": "last_agent_id", "type": "STRING"},
+        {"name": "last_user_id", "type": "INTEGER"},
+        {"name": "call_survey", "type": "STRING"},
+        {"name": "call_survey_result", "type": "INTEGER"},
+        {"name": "missed_reason", "type": "STRING"},
+    ],
+    id_key="id",
+    cursor_key="start_time",
+)
