+++ pyproject.toml
@@ -11,6 +11,8 @@ aiohttp = "^3.7.4"
 google-cloud-bigquery = "^2.27.1"
 google-cloud-tasks = "^2.5.2"
 asyncio-throttle = "^1.0.2"
+httpx = "^0.22.0"
+compose = "^1.2.8"
 
 [tool.poetry.dev-dependencies]
 pytest = "^6.2.5"
